05-10-2024

+++ src/main/java/com/zayaanit/controller/FA15.java
@@ -1,5 +1,6 @@
 package com.zayaanit.controller;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -20,19 +21,29 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import com.zayaanit.entity.Acdetail;
 import com.zayaanit.entity.Acheader;
+import com.zayaanit.entity.Acmst;
 import com.zayaanit.entity.Acsub;
 import com.zayaanit.entity.Cabunit;
+import com.zayaanit.entity.Cadoc;
 import com.zayaanit.entity.Xscreens;
+import com.zayaanit.entity.pk.AcdetailPK;
 import com.zayaanit.entity.pk.AcheaderPK;
+import com.zayaanit.entity.pk.AcmstPK;
 import com.zayaanit.entity.pk.AcsubPK;
 import com.zayaanit.entity.pk.CabunitPK;
 import com.zayaanit.entity.pk.XscreensPK;
 import com.zayaanit.enums.SubmitFor;
 import com.zayaanit.model.ReloadSection;
+import com.zayaanit.model.YearPeriodResult;
+import com.zayaanit.repository.AcdetailRepo;
 import com.zayaanit.repository.AcheaderRepo;
+import com.zayaanit.repository.AcmstRepo;
 import com.zayaanit.repository.AcsubRepo;
 import com.zayaanit.repository.CabunitRepo;
+import com.zayaanit.repository.CadocRepo;
+import com.zayaanit.service.AcheaderService;
 
 /**
  * @author Zubayer Ahamed
@@ -44,7 +55,11 @@ public class FA15 extends KitController {
 
 	@Autowired private AcheaderRepo acheaderRepo;
 	@Autowired private CabunitRepo cabunitRepo;
+	@Autowired private AcmstRepo acmstRepo;
 	@Autowired private AcsubRepo acsubRepo;
+	@Autowired private AcheaderService acheaderService;
+	@Autowired private AcdetailRepo acdetailRepo;
+	@Autowired private CadocRepo cadocRepo;
 
 	private String pageTitle = null;
 
@@ -88,11 +103,72 @@ public class FA15 extends KitController {
 				}
 			}
 			model.addAttribute("acheader", acheader != null ? acheader : Acheader.getDefaultInstance());
+
+			List<Cadoc> cdocList = cadocRepo.findAllByZidAndXscreenAndXtrnnum(sessionManager.getBusinessId(), "FA15", Integer.valueOf(xvoucher));
+			model.addAttribute("documents", cdocList);
+
 			return "pages/FA15/FA15-fragments::main-form";
 		}
 
 		model.addAttribute("acheader", Acheader.getDefaultInstance());
 		return "pages/FA15/FA15";
+	}
+
+	@GetMapping("/detail-table")
+	public String detailFormFragment(@RequestParam String xvoucher, @RequestParam String xrow, @RequestParam(required = false) Integer xacc, Model model) {
+		if("RESET".equalsIgnoreCase(xvoucher) && "RESET".equalsIgnoreCase(xrow)) {
+			return "pages/FA15/FA15-fragments::detail-table";
+		}
+
+		Optional<Acheader> oph = acheaderRepo.findById(new AcheaderPK(sessionManager.getBusinessId(), Integer.parseInt(xvoucher)));
+		if(!oph.isPresent()) return "pages/FA15/FA15-fragments::detail-table";
+		model.addAttribute("acheader", oph.get());
+
+		List<Acdetail> detailList = acdetailRepo.findAllByZidAndXvoucher(sessionManager.getBusinessId(), Integer.parseInt(xvoucher));
+		for(Acdetail detail : detailList) {
+			Optional<Acmst> accountOp =  acmstRepo.findById(new AcmstPK(sessionManager.getBusinessId(), detail.getXacc()));
+			if(accountOp.isPresent()) {
+				detail.setAccountName(accountOp.get().getXdesc());
+				detail.setAccountUsage(accountOp.get().getXaccusage());
+			}
+
+			Optional<Acsub> subAccountOp = acsubRepo.findById(new AcsubPK(sessionManager.getBusinessId(), detail.getXsub()));
+			if(subAccountOp.isPresent()) detail.setSubAccountName(subAccountOp.get().getXname());
+		}
+		model.addAttribute("detailList", detailList);
+
+		Acmst account = null;
+		if(xacc != null) {
+			Optional<Acmst> accountOp =  acmstRepo.findById(new AcmstPK(sessionManager.getBusinessId(), xacc));
+			account = accountOp.isPresent() ? accountOp.get() : null;
+		}
+
+		if("RESET".equalsIgnoreCase(xrow)) {
+			Acdetail acdetail = Acdetail.getDefaultInstance(Integer.parseInt(xvoucher));
+			if(account != null) {
+				acdetail.setXacc(account.getXacc());
+				acdetail.setAccountName(account.getXdesc());
+				acdetail.setAccountUsage(account.getXaccusage());
+			}
+
+			model.addAttribute("acdetail", acdetail);
+			return "pages/FA15/FA15-fragments::detail-table";
+		}
+
+		Optional<Acdetail> acdetailOp = acdetailRepo.findById(new AcdetailPK(sessionManager.getBusinessId(), Integer.parseInt(xvoucher), Integer.parseInt(xrow)));
+		Acdetail acdetail = acdetailOp.isPresent() ? acdetailOp.get() : Acdetail.getDefaultInstance(Integer.parseInt(xvoucher));
+		if(acdetail != null && acdetail.getXacc() != null) {
+			Optional<Acmst> accountOp =  acmstRepo.findById(new AcmstPK(sessionManager.getBusinessId(), acdetail.getXacc()));
+			account = accountOp.isPresent() ? accountOp.get() : null;
+		}
+		if(account != null) {
+			acdetail.setXacc(account.getXacc());
+			acdetail.setAccountName(account.getXdesc());
+			acdetail.setAccountUsage(account.getXaccusage());
+		}
+
+		model.addAttribute("acdetail", acdetail);
+		return "pages/FA15/FA15-fragments::detail-table";
 	}
 
 	@PostMapping("/store")
@@ -102,22 +178,48 @@ public class FA15 extends KitController {
 		modelValidator.validateAcheader(acheader, bindingResult, validator);
 		if(bindingResult.hasErrors()) return modelValidator.getValidationMessage(bindingResult);
 
-
-		if(StringUtils.isBlank(acheader.getXtype())) {
-			responseHelper.setErrorStatusAndMessage("Account Type required");
+		if(acheader.getXdate() == null) {
+			responseHelper.setErrorStatusAndMessage("Voucher date required");
 			return responseHelper.getResponse();
 		}
 
+		if(acheader.getXbuid() == null) {
+			responseHelper.setErrorStatusAndMessage("Business unit required");
+			return responseHelper.getResponse();
+		}
+
+		if(StringUtils.isBlank(acheader.getXvtype())) {
+			responseHelper.setErrorStatusAndMessage("Voucher Type required");
+			return responseHelper.getResponse();
+		}
+
+		YearPeriodResult yp = acheaderService.getYearPeriod(acheader.getXdate());
+		if(yp == null) {
+			responseHelper.setErrorStatusAndMessage("Error with voucher year period.");
+			return responseHelper.getResponse();
+		}
+		acheader.setXyear(yp.getYear());
+		acheader.setXper(yp.getPeriod());
+
+		if(sessionManager.getLoggedInUserDetails().getXstaff() == null) {
+			responseHelper.setErrorStatusAndMessage("Staff not set with this user");
+			return responseHelper.getResponse();
+		}
+		acheader.setXstaff(sessionManager.getLoggedInUserDetails().getXstaff());
+
 		// Create new
 		if(SubmitFor.INSERT.equals(acheader.getSubmitFor())) {
+			acheader.setXtype("General");
+			acheader.setXstatusjv("Balanced");
 			acheader.setXvoucher(xscreenRepo.Fn_getTrn(sessionManager.getBusinessId(), "FA15"));
 			acheader.setZid(sessionManager.getBusinessId());
 			acheader = acheaderRepo.save(acheader);
 
 			List<ReloadSection> reloadSections = new ArrayList<>();
-			reloadSections.add(new ReloadSection("main-form-container", "/FA15?xvoucher=RESET"));
+			reloadSections.add(new ReloadSection("main-form-container", "/FA15?xvoucher=" + acheader.getXvoucher()));
+			reloadSections.add(new ReloadSection("detail-table-container", "/FA15/detail-table?xvoucher="+ acheader.getXvoucher() +"&xrow=RESET"));
 			responseHelper.setReloadSections(reloadSections);
-			responseHelper.setSuccessStatusAndMessage("Saved successfully");
+			responseHelper.setSuccessStatusAndMessage("Voucher created successfully");
 			return responseHelper.getResponse();
 		}
 
@@ -128,15 +230,117 @@ public class FA15 extends KitController {
 			return responseHelper.getResponse();
 		}
 
+		if("Posted".equalsIgnoreCase(acheader.getXstatusjv())) {
+			responseHelper.setErrorStatusAndMessage("Voucher already Posted");
+			return responseHelper.getResponse();
+		}
+
 		Acheader existObj = op.get();
-		BeanUtils.copyProperties(acheader, existObj, "zid", "zuserid", "ztime", "xvoucher");
+		BeanUtils.copyProperties(acheader, existObj, "zid", "zuserid", "ztime", "xvoucher", "xtype", "xstatusjv");
 
 		existObj = acheaderRepo.save(existObj);
 
 		List<ReloadSection> reloadSections = new ArrayList<>();
 		reloadSections.add(new ReloadSection("main-form-container", "/FA15?xvoucher=" + existObj.getXvoucher()));
+		reloadSections.add(new ReloadSection("detail-table-container", "/FA15/detail-table?xvoucher="+ acheader.getXvoucher() +"&xrow=RESET"));
 		responseHelper.setReloadSections(reloadSections);
-		responseHelper.setSuccessStatusAndMessage("Updated successfully");
+		responseHelper.setSuccessStatusAndMessage("Voucher updated successfully");
+		return responseHelper.getResponse();
+	}
+
+	@PostMapping("/detail/store")
+	public @ResponseBody Map<String, Object> storeDetail(Acdetail acdetail, BindingResult bindingResult){
+		if(acdetail.getXvoucher() == null) {
+			responseHelper.setErrorStatusAndMessage("Voucher not found");
+			return responseHelper.getResponse();
+		}
+
+		Optional<Acheader> oph = acheaderRepo.findById(new AcheaderPK(sessionManager.getBusinessId(), acdetail.getXvoucher()));
+		if(!oph.isPresent()) {
+			responseHelper.setErrorStatusAndMessage("Voucher not found");
+			return responseHelper.getResponse();
+		}
+
+		Acheader acheader = oph.get();
+		if("Posted".equals(acheader.getXstatusjv())) {
+			responseHelper.setErrorStatusAndMessage("Voucher already posted");
+			return responseHelper.getResponse();
+		}
+
+		if(acdetail.getXacc() == null) {
+			responseHelper.setErrorStatusAndMessage("Account requried");
+			return responseHelper.getResponse();
+		}
+
+		Optional<Acmst> accountOp =  acmstRepo.findById(new AcmstPK(sessionManager.getBusinessId(), acdetail.getXacc()));
+		if(!accountOp.isPresent()) {
+			responseHelper.setErrorStatusAndMessage("Account not found");
+			return responseHelper.getResponse();
+		}
+		Acmst account = accountOp.get();
+		if(!"Default".equalsIgnoreCase(account.getXaccusage())) {
+			if(acdetail.getXsub() == null) {
+				responseHelper.setErrorStatusAndMessage("Sub account required");
+				return responseHelper.getResponse();
+			}
+		}
+
+		if(acdetail.getXdebit() == null) acdetail.setXdebit(BigDecimal.ZERO);
+		if(acdetail.getXcredit() == null) acdetail.setXcredit(BigDecimal.ZERO);
+
+		if(acdetail.getXdebit().compareTo(BigDecimal.ZERO) == 1 && acdetail.getXcredit().compareTo(BigDecimal.ZERO) == 1) {
+			responseHelper.setErrorStatusAndMessage("Enter only debit or credit amount!");
+			return responseHelper.getResponse();
+		}
+
+		if(acdetail.getXdebit().compareTo(BigDecimal.ZERO) == 0 && acdetail.getXcredit().compareTo(BigDecimal.ZERO) == 0) {
+			responseHelper.setErrorStatusAndMessage("Enter valid amount!");
+			return responseHelper.getResponse();
+		}
+
+		acdetail.setXprime(acdetail.getXdebit().subtract(acdetail.getXcredit()));
+
+		// Create new
+		if(SubmitFor.INSERT.equals(acdetail.getSubmitFor())) {
+			acdetail.setXrow(acdetailRepo.getNextAvailableRow(sessionManager.getBusinessId(), acdetail.getXvoucher()));
+			acdetail.setZid(sessionManager.getBusinessId());
+			acdetail = acdetailRepo.save(acdetail);
+
+			BigDecimal total = acdetailRepo.getTotalPrimeAmount(sessionManager.getBusinessId(), acdetail.getXvoucher());
+			acheader.setXstatusjv(total.compareTo(BigDecimal.ZERO) == 0 ? "Balanced" : "Suspended");
+			acheaderRepo.save(acheader);
+
+			List<ReloadSection> reloadSections = new ArrayList<>();
+			reloadSections.add(new ReloadSection("main-form-container", "/FA15?xvoucher=" + acdetail.getXvoucher()));
+			reloadSections.add(new ReloadSection("detail-table-container", "/FA15/detail-table?xvoucher=" + acdetail.getXvoucher() + "&xrow=RESET"));
+			responseHelper.setReloadSections(reloadSections);
+			responseHelper.setSuccessStatusAndMessage("Voucher detail added successfully");
+			return responseHelper.getResponse();
+		}
+
+		Optional<Acdetail> existOp = acdetailRepo.findById(new AcdetailPK(sessionManager.getBusinessId(), acdetail.getXvoucher(), acdetail.getXrow()));
+		if(!existOp.isPresent()) {
+			responseHelper.setErrorStatusAndMessage("Detail not found in this system");
+			return responseHelper.getResponse();
+		}
+
+		Acdetail exist = existOp.get();
+		BeanUtils.copyProperties(acdetail, exist, "zid", "zuserid", "ztime", "xvoucher", "xrow", "xacc");
+		exist = acdetailRepo.save(exist);
+		if(exist == null) {
+			responseHelper.setErrorStatusAndMessage("Update failed");
+			return responseHelper.getResponse();
+		}
+
+		BigDecimal total = acdetailRepo.getTotalPrimeAmount(sessionManager.getBusinessId(), exist.getXvoucher());
+		acheader.setXstatusjv(total.compareTo(BigDecimal.ZERO) == 0 ? "Balanced" : "Suspended");
+		acheaderRepo.save(acheader);
+
+		List<ReloadSection> reloadSections = new ArrayList<>();
+		reloadSections.add(new ReloadSection("main-form-container", "/FA15?xvoucher=" + acdetail.getXvoucher()));
+		reloadSections.add(new ReloadSection("detail-table-container", "/FA15/detail-table?xvoucher=" + acdetail.getXvoucher() + "&xrow=" + exist.getXrow()));
+		responseHelper.setReloadSections(reloadSections);
+		responseHelper.setSuccessStatusAndMessage("Voucher detail updated successfully");
 		return responseHelper.getResponse();
 	}
 
@@ -148,11 +352,56 @@ public class FA15 extends KitController {
 			return responseHelper.getResponse();
 		}
 
+		if(op.get().getXstatusjv().equals("Posted")) {
+			responseHelper.setErrorStatusAndMessage("Voucher already posted");
+			return responseHelper.getResponse();
+		}
+
+		acdetailRepo.deleteAllByZidAndXvoucher(sessionManager.getBusinessId(), xvoucher);
+
 		Acheader obj = op.get();
 		acheaderRepo.delete(obj);
 
 		List<ReloadSection> reloadSections = new ArrayList<>();
 		reloadSections.add(new ReloadSection("main-form-container", "/FA15?xvoucher=RESET"));
+		reloadSections.add(new ReloadSection("detail-table-container", "/FA15/detail-table?xvoucher=RESET&xrow=RESET"));
+		responseHelper.setReloadSections(reloadSections);
+		responseHelper.setSuccessStatusAndMessage("Deleted successfully");
+		return responseHelper.getResponse();
+	}
+
+	@DeleteMapping("/detail-table")
+	public @ResponseBody Map<String, Object> deleteDetail(@RequestParam Integer xvoucher, @RequestParam Integer xrow){
+		Optional<Acheader> oph = acheaderRepo.findById(new AcheaderPK(sessionManager.getBusinessId(), xvoucher));
+		if(!oph.isPresent()) {
+			responseHelper.setErrorStatusAndMessage("Voucher not found");
+			return responseHelper.getResponse();
+		}
+
+		Acheader acheader = oph.get();
+
+		if("Posted".equals(acheader.getXstatusjv())) {
+			responseHelper.setErrorStatusAndMessage("Voucher already posted");
+			return responseHelper.getResponse();
+		}
+
+		Optional<Acdetail> op = acdetailRepo.findById(new AcdetailPK(sessionManager.getBusinessId(), xvoucher, xrow));
+		if(!op.isPresent()) {
+			responseHelper.setErrorStatusAndMessage("Detail not found");
+			return responseHelper.getResponse();
+		}
+
+		Acdetail obj = op.get();
+		acdetailRepo.delete(obj);
+
+		// Update line amount and total amount of header
+		BigDecimal total = acdetailRepo.getTotalPrimeAmount(sessionManager.getBusinessId(), xvoucher);
+		acheader.setXstatusjv(total.compareTo(BigDecimal.ZERO) == 0 ? "Balanced" : "Suspended");
+		acheaderRepo.save(acheader);
+
+		List<ReloadSection> reloadSections = new ArrayList<>();
+		reloadSections.add(new ReloadSection("main-form-container", "/FA15?xvoucher=" + xvoucher));
+		reloadSections.add(new ReloadSection("detail-table-container", "/FA15/detail-table?xvoucher="+xvoucher+"&xrow=RESET"));
 		responseHelper.setReloadSections(reloadSections);
 		responseHelper.setSuccessStatusAndMessage("Deleted successfully");
 		return responseHelper.getResponse();
